using OS version of the 'unzip' command to unzip (instead of 'unzip' npm module)
