+++ package.json
@@ -21,7 +21,10 @@
   "license": "MIT",
   "dependencies": {
     "browserify": "^5.0.0",
+    "city-blocks": "git://github.com/anthill/city-blocks.git",
+    "commander": "^2.3.0",
     "compression": "^1.1.0",
+    "csv-parse": "0.0.6",
     "es6-map": "^0.1.0",
     "es6-promise": "^1.0.0",
     "es6-promisify": "^1.1.1",
@@ -31,18 +34,13 @@
     "graceful-fs": "^3.0.2",
     "jsdom": "^1.0.1",
     "rbush": "^1.3.2",
+    "rimraf": "^2.2.8",
     "socket.io": "^1.0.6",
     "socket.io-client": "^1.0.6",
     "three": "^0.68.0",
-    "csv-parse": "0.0.6",
-    "city-blocks": "git://github.com/iOiurson/city-blocks.git#raycasting"
+    "tmp": "^0.0.24"
   },
   "devDependencies": {
-    "unzip": "^0.1.9",
-    "commander": "^2.3.0",
-    "tmp": "^0.0.24",
-    "q": "^1.0.1",
-    "glob": "^4.0.5",
     "watchify": "^1.0.2"
   }
 }
