+++ tools/unzipCUB3ds.js
@@ -1,40 +1,33 @@
 "use strict";
 
+var Map = require('es6-map');
+var Promise = require('es6-promise').Promise;
+
 var path = require('path');
+var spawn = require('child_process').spawn;
 var fs = require('graceful-fs');
 
 var program = require('commander');
-var unzip = require('unzip');
 var tmp = require('tmp');
-var Q = require('q');
-var Map = require('es6-map');
+var promisify = require("es6-promisify");
 
-var containingCube = require('./containingCube.js');
-var computeMeshVolume = require('./computeMeshVolume.js');
-var _3dsFormatToAntsBinaryBuffer = require('./3dsFormatToAntsBinaryBuffer.js');
+var _3dsObjectToBuildingBuffersAndMetadata = require('./_3dsObjectToBuildingBuffersAndMetadata');
 var parse3ds = require('./parse3ds.js');
 
 var getTileAltitudes = require('./getTileAltitudes.js');
 var tileAltitudesP = getTileAltitudes(path.resolve(__dirname, '../data/DALLAGE_3D.csv'));
 
-
-
 function tmpdir(){
-    var def = Q.defer();
-
-    tmp.dir(function(err, dir){
-        if(err)
-            def.reject(err);
-        else
-            def.resolve(dir);
+    return new Promise(function(resolve, reject){
+        tmp.dir(function(err, dir){
+            if(err) reject(err); else resolve(dir);
+        });
     });
-
-    return def.promise;
 }
-var readdir = Q.nfbind(fs.readdir);
-var readFile = Q.nfbind(fs.readFile);
-var writeFile = Q.nfbind(fs.writeFile);
-var lstat = Q.nfbind(fs.lstat);
+var readdir = promisify(fs.readdir);
+var readFile = promisify(fs.readFile);
+var writeFile = promisify(fs.writeFile);
+var lstat = promisify(fs.lstat);
 
 
 /*
@@ -43,9 +36,7 @@ var lstat = Q.nfbind(fs.lstat);
 */
 function allInSequence(arr, f){
     if(arr.length === 0){
-        var def = Q.defer();
-        def.resolve([]);
-        return def.promise;
+        return Promise.resolve([]);
     }
 
     if(arr.length === 1){
@@ -83,151 +74,25 @@ var outAbsolutePath = path.resolve(process.cwd(), program.out);
 console.log(zipAbsolutePath, outAbsolutePath);
 
 
-// name => next nb to use to append to the name to make it unique
-var ids = new Map();
-
-// in : one archive
-// out : a bunch of binaries + metadata
-
 function extractBuildings(_3dsPath, x, y){
-    var def = Q.defer();
-    parse3ds(_3dsPath, function(err, data){
-        if(err){
-            def.reject(err);
-            return;
-        }
-
-        var objects = data.getObjects();
-        var meshes = objects.map(function(o){
-            // Generate unique ids since the data producers may fail to do so
-            var id;
-            if(!ids.has(o.name)){
-                id = o.name;
-                ids.set(o.name, 1);
+    //console.log('extracting buildings from', _3dsPath);
+    
+    return new Promise(function(resolve, reject){
+        parse3ds(_3dsPath, function(err, data){
+            if(err){
+                reject(err);
+                return;
             }
-            else{
-                var nb = ids.get(o.name);
-                id = o.name + '-' + nb;
-                while(ids.has(id)){
-                    nb++;
-                    id = o.name + '-' + nb;
-                }
-                ids.set(o.name, nb+1);
-            }
-
-            return {
-                id: id,
-                vertices: o.meshes.vertices,
-                faces: o.meshes.faces
-            };
-
-        });
-
-        /*
-            Working around https://github.com/anthill/bordeaux3d/issues/11
-        */
-        // looking for xXXXyYYY object. There is only one per 3ds file
-        var xyObject = meshes.filter(function(m){ return !!m.id.match(/x(\d{1,4})y(\d{1,4})/) })[0];
-        var xyContainingCube = containingCube(xyObject);
-
-        // Ideally, xyContainingCube.xmax and xyContainingCube.ymax should be +100. Finding the translation.
-        var deltaX = 100 - xyContainingCube.maxX;
-        var deltaY = 100 - xyContainingCube.maxY;
-
-        if(deltaX !== 0 || deltaY !== 0){
-            // apply translation to all tile objects
-            meshes.forEach(function(m){
-                m.vertices.forEach(function(v){
-                    v.x += deltaX;
-                    v.y += deltaY;
-                });
-            });
-        }
-
-
-        tileAltitudesP.then(function(tilesAltitudes){
+            
             var key = _3dsPath.match(/x\d{1,4}y\d{1,4}/)[0];
-
-            var deltaZ = tilesAltitudes[key];
-
-            meshes.forEach(function(m){
-                m.vertices.forEach(function(v){
-                    v.z += deltaZ;
-                });
-            });
             
-            // Find tile bounding box
-            var containingCubes = meshes.map(containingCube);
-            // create a fake mesh based on the cubes descriptions
-            var fakeCombiningMesh = {
-                vertices: containingCubes.reduce(function(acc, cubeDesc){
-                    acc.push({
-                        x: cubeDesc.minX,
-                        y: cubeDesc.minY,
-                        z: cubeDesc.minZ,
-                    });
-                    acc.push({
-                        x: cubeDesc.maxX,
-                        y: cubeDesc.maxY,
-                        z: cubeDesc.maxZ,
-                    });
-
-                    return acc;
-                }, [])
-            };
-            var tileContainingCube = containingCube(fakeCombiningMesh);
-
-            var minZ = Math.floor(tileContainingCube.minZ);
-            var maxZ = Math.ceil(tileContainingCube.maxZ);
-            if(minZ === maxZ){ // happens for flat floor objects
-                maxZ = minZ + 1;
-            }
-
-            // integer approximation
-            var tileMetadata = {
-                X: x,
-                Y: y,
-                minX: Math.floor(tileContainingCube.minX),
-                maxX: Math.ceil(tileContainingCube.maxX),
-                minY: Math.floor(tileContainingCube.minY),
-                maxY: Math.ceil(tileContainingCube.maxY),
-                minZ: minZ,
-                maxZ: maxZ,
-                objects: Object.create(null)
-            };
-
-            var buildingBuffers = Object.create(null);
-            meshes.forEach(function(m, i){
-                try{
-                buildingBuffers[m.id] = _3dsFormatToAntsBinaryBuffer(m, tileMetadata);
-                }
-                catch(e){
-                    console.error('compacting error', x, y, e)
-                }
-            });
-
-            meshes.forEach(function(m){
-                // for recentering
-                var objectCube = containingCube(m);
-
-                tileMetadata.objects[m.id] = {
-                    x: Math.round( (objectCube.minX + objectCube.maxX)/2 ),
-                    y: Math.round( (objectCube.minY + objectCube.maxY)/2 ),
-                };
-            });
-
-            def.resolve({
-                buildingBuffers: buildingBuffers,
-                tileMetadata : tileMetadata
-            });
-            
-            
+            resolve(tileAltitudesP.then(function(tilesAltitudes){
+                var deltaZ = tilesAltitudes[key];
+                
+                return _3dsObjectToBuildingBuffersAndMetadata(data, x, y, deltaZ);
+            }));
         });
-        
-
     });
-
-    return def.promise;
 }
 
 
@@ -235,26 +100,17 @@ function unzipInTmpDir(pathToZip){
     var readStream = fs.createReadStream(pathToZip);
 
     return tmpdir().then(function(tmpDir){
-        var def = Q.defer();
-
-        var extractWriteStream = unzip.Extract({ path: tmpDir })
-        readStream.pipe(extractWriteStream);
-
-        extractWriteStream.on('close', function(e){ // 'close' event, unlike 'finish' guarantees all writes to disk are finished
-            //console.log('close');
-            def.resolve(tmpDir);
+        return new Promise(function(resolve, reject){
+            var args = [pathToZip, '-d', tmpDir];
+            
+            //console.log('unzip', args.join(' '));
+            
+            var unzipProc = spawn('unzip', args);
+            
+            unzipProc.on('exit', function(code, signal){
+                resolve(tmpDir);
+            });
         });
-
-        extractWriteStream.on('error', function(err){
-            def.reject(err);
-        });
-
-        extractWriteStream.on('finish', function(e){ // 'close' event, unlike 'finish' guarantees all writes to disk are finished
-            console.log('finish', pathToZip, tmpDir);
-            extractWriteStream.end();
-        });
-
-        return def.promise;
     });
 }
 
@@ -263,7 +119,7 @@ function unzipInTmpDir(pathToZip){
     Process the directory of a single selection
 */
 function processSelectionDirectory(selectionZipDirPath){
-    //console.log('processSelectionDirectory', selectionZipDirPath);
+    console.log('processing', selectionZipDirPath);
 
     var selectionName = path.basename(selectionZipDirPath, '.zip'); // assumed 2 letters like "RL"
 
@@ -293,7 +149,7 @@ function processSelectionDirectory(selectionZipDirPath){
                         });
                     //console.log(name, tile3dsPaths);
 
-                    return Q.all(tile3dsPathsAndxy.map(function(tpxy){
+                    return Promise.all(tile3dsPathsAndxy.map(function(tpxy){
                         var tile3dsPath = tpxy.path;
                         var x = tpxy.x;
                         var y = tpxy.y;
@@ -302,22 +158,19 @@ function processSelectionDirectory(selectionZipDirPath){
                             var buildingBuffers = res.buildingBuffers;
                             var tileMetadata = res.tileMetadata;
 
-                            return Q.all(Object.keys(buildingBuffers).map(function(id){
+                            return Promise.all(Object.keys(buildingBuffers).map(function(id){
                                 var buildingOutPath = path.join(outAbsolutePath, id);
                                 
                                 // despite graceful-fs, we see some EMFILE errors in writes
                                 function tryWriteFile(){
-                                    return writeFile(buildingOutPath, buildingBuffers[id]).fail(function(err){
+                                    return writeFile(buildingOutPath, buildingBuffers[id]).catch(function(err){
                                         if(String(err).indexOf('EMFILE') !== -1){
-                                            console.error('tryWriteFile EMFILE', selectionName, x, y, tile3dsPath, err)
-                                            var def = Q.defer();
-                                            
-                                            // retry later
-                                            setTimeout(function(){
-                                                def.resolve(tryWriteFile());
-                                            }, 100);
-
-                                            return def.promise;
+                                            console.error('tryWriteFile EMFILE', selectionName, x, y, tile3dsPath, err);
+                                            return new Promise(function(resolve){
+                                                setTimeout(function(){
+                                                    resolve(tryWriteFile());
+                                                }, 100);
+                                            });
                                         }
                                         else{// forward error
                                             throw err;
@@ -331,7 +184,7 @@ function processSelectionDirectory(selectionZipDirPath){
                                 return tileMetadata;
                             });
 
-                        }).fail(function(err){
+                        }).catch(function(err){
                             console.error('extractBuildings error', selectionName, x, y, tile3dsPath, err);
                         });
                     }));
@@ -349,26 +202,17 @@ console.time('extract');
 unzipInTmpDir(zipAbsolutePath)
     .then(function(tmpDir){
         console.timeEnd('extract');
-
+    
         return readdir(tmpDir).then(function(selectionZips){
-            //console.log("selectionZips", selectionZips);
-
-            //selectionZips = selectionZips.slice(0, 3);
 
             var absoluteZipPaths = selectionZips.map(function(zipPath){
                 return path.join(tmpDir, zipPath);
             });
 
-            //console.log(absoluteZipPaths);
-
-            //return Q.all(absoluteZipPaths.map(processSelectionDirectory));
-
             return allInSequence(absoluteZipPaths, processSelectionDirectory);
         });
     })
     .then(function(dallesMetadata){
-        //console.log('final result', dallesMetadata);
-
         var tilesMetadata = dallesMetadata.reduce(function(acc, tm){
             return acc.concat(tm)
         }, [])
@@ -382,12 +226,8 @@ unzipInTmpDir(zipAbsolutePath)
         return writeFile(path.join(outAbsolutePath, 'metadata.json'), JSON.stringify(tilesMetadata));
 
     })
-    .then(function(){
-
-        console.timeEnd('all');
-
-    })
-    .fail(function(err){ console.error(err) });
+    .then(function(){ console.timeEnd('all'); })
+    .catch(function(err){ console.error(err) });
